pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "platito"
version = "0.1.0"
description = "Protein Language-aware Transferable Implicit Transfer Operator for coarse-grained molecular dynamics"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # PyTorch (CPU) — install with [cuda] extra for GPU support (see README)
    "torch==2.7.1",
    "torchdiffeq==0.2.5",
    # Training framework
    "lightning==2.5.1",
    "hydra-core==1.3.2",
    "hydra-colorlog",
    "omegaconf==2.3.0",
    "wandb==0.19.0",
    # Protein / MD
    "mdtraj==1.10.3",
    "biopython==1.85",
    "deeptime",
    "h5py==3.13",
    # ESM Cambrian — for precomputing pLM embeddings
    "esm==3.2.0",
    # Utilities
    "rich",
    "tqdm",
    "einops==0.8",
    "huggingface_hub",
    "matplotlib==3.9.3",
    "nglview==3.1.4",
    "notebook",
    "ipympl",
]

[project.optional-dependencies]
cuda = ["torch==2.7.1+cu126"]

[tool.setuptools.packages.find]
where = ["."]
include = ["platito*"]
